use long day string

+++ src/lib/helpers.ts
@@ -1,7 +1,7 @@
 export const generateCustomDate = (): CustomDate => {
   const dateObj = new Date();
   return {
-    day: dateObj.toLocaleDateString('en-US', { weekday: 'short' }),
+    day: dateObj.toLocaleDateString('en-US', { weekday: 'long' }),
     hour: dateObj.toLocaleTimeString([], { hour: 'numeric' }),
   };
 };
